Delete sent messages
If the publisher are subscriber are both online and the publisher sends
a message to the subscriber, the message is not deleted from the
backend store (found in both redid and mongo).

The reason was different messageIds were used for storing the packet in
server.js's publish method and updating the offline packet in
client.js's this.forward method.

Updated the publish method in server.js to include the messageId
generated for storing (for newPacket) and use it in the this.forward
method of client.js to update the packet using updateOfflinePacket
method.